+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## main
 
+### ✨ Features and improvements
+
+- Add getLayersOrder() to Map and Style ([#3279](https://github.com/maplibre/maplibre-gl-js/pull/3279))
 - _...Add new stuff here..._
 
 ### 🐞 Bug fixes